消息格式调整、消息编解码修改、分发器调整

+++ gcore/request.go
@@ -24,6 +24,6 @@ func (r *Request) ConnID() uint64 {
 }
 
 // ID 消息ID
-func (r *Request) ID() uint16 {
+func (r *Request) ID() uint32 {
 	return r.Message.ID()
 }
