+++ trait/message.go
@@ -1,11 +1,11 @@
 package trait
 
 type Message interface {
-	ID() uint16
-	DataLen() uint16
+	ID() uint32
+	DataLen() uint32
 	Data() []byte
 
-	SetID(uint16)
-	SetDataLen(uint16)
+	SetID(uint32)
+	SetDataLen(uint32)
 	SetData([]byte)
 }
